ml_service/requirements.txt
fastapi
uvicorn
python-multipart
pydantic
joblib
pandas
Pillow
torch
torchvision
catboost
scikit-learn